requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
autogen==1.0.16
autopep8==2.1.2
backports.tarfile==1.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docutils==0.21.2
dotenv-python==0.0.1
executing==2.0.1
filelock==3.14.0
FLAML==2.1.2
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==25.2.1
langchain-core==0.2.5
langchain-openai==0.1.8
langsmith==0.1.77
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mlx==0.15.0
more-itertools==10.2.0
neo4j==5.20.0
neo4j-genai==0.2.0
nest-asyncio==1.6.0
nh3==0.2.17
numpy==1.26.4
openai==1.30.4
orjson==3.10.4
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyautogen==0.2.27
pycodestyle==2.11.1
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
readme_renderer==43.0
regex==2024.5.15
requests==2.32.2
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
twine==5.1.0
types-mock==5.1.0.20240425
types-requests==2.32.0.20240602
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.19.0
